requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
openai
google-genai>=1.32.0
tiktoken>=0.5.0
regex>=2022.1.18
httpx==0.24.1
python-dotenv==1.0.0
python-multipart==0.0.6
urllib3==1.26.18
requests-toolbelt==0.9.1
supabase==1.0.4